booking-system/requirements.txt
Django>=4.2.0,<4.3.0
djangorestframework>=3.14.0,<3.15.0
Pillow>=10.0.0,<10.1.0
django-filter>=23.2,<24.0
gunicorn==23.0.0
psycopg2-binary==2.9.10